poke into submission
