+++ movienightbot/db/controllers.py
@@ -9,7 +9,7 @@ import peewee as pw
 
 from ..exc import VoteError
 from . import BaseController
-from ..util import get_imdb_info_by_id
+from ..imdb import get_imdb_info_by_id
 from .models import (
     IMDBInfo,
     Movie,
@@ -38,13 +38,13 @@ class IMDBInfoController(BaseController):
         imdb_data = {
             "imdb_id": found_imdb_id,
             "title": imdb_info.title,
-            "canonical_title": imdb_info.canonical_title if imdb_info.canonical_title else imdb_info.title,
+            "canonical_title": imdb_info.title_localized if imdb_info.title_localized else imdb_info.title,
             "year": imdb_info.year if imdb_info.year else 0,
             "thumbnail_poster_url": imdb_info.cover_url if imdb_info.cover_url else "",
-            "full_size_poster_url": imdb_info.full_size_url if imdb_info.full_size_url else "",
+            "full_size_poster_url": imdb_info.cover_url if imdb_info.cover_url else "",
         }
         imdb_controller = IMDBInfoController()
-        return self.create(imdb_data)
+        return imdb_controller.create(imdb_data)
 
     def get_by_id(self, imdb_id: str) -> Union[Vote, None]:
         return super().get_by_id(id=imdb_id, primary_key="imdb_id")
