+++ movienightbot/util.py
@@ -1,11 +1,9 @@
 import asyncio
 import datetime
 import logging
-import re
-from typing import Optional, Union
+from typing import Union
 
 import discord
-import imdbinfo
 import peewee as pw
 
 from .db.controllers import MovieVote, MovieVoteController, ServerController
@@ -147,52 +145,10 @@ emojis_text = {
 emojis_unicode = {v: k for k, v in emojis_text.items()}
 
 
-imdb_url_regex = re.compile(r"title/tt([0-9]+)")
-
-
 async def add_vote_emojis(vote_msg: discord.Message, movie_votes: MovieVote):
     for movie_vote in movie_votes:
         await vote_msg.add_reaction(emojis_text[movie_vote.emoji])
     await vote_msg.add_reaction(emojis_text[":arrows_counterclockwise:"])
-
-def get_imdb_info_by_id(imdb_id: Union[int, str]) -> Union[None, imdbinfo.services.MovieDetail]:
-    if not imdb_id:
-        return None
-
-    return imdbinfo.get_movie(str(imdb_id))
-
-
-def get_imdb_info(movie_name: str, kind: Optional[str] = None, year: Optional[int] = None) -> Union[None, imdbinfo.services.MovieDetail]:
-    if not movie_name:
-        return None
-
-    if movie_name.lower().startswith("http"):
-        movie_id = imdb_url_regex.findall(movie_name)
-        logger.debug("movie regex: `{}` >> {}", movie_name, movie_id)
-        if len(movie_id) == 1:
-            imdb_id = movie_id[0]
-            return_movie = get_imdb_info_by_id(imdb_id)
-        else:
-            return None
-    else:
-        logger.debug(f"searching for `{movie_name}`")
-        results = imdbinfo.search_title(movie_name)
-        logger.debug("IMDB RESULTS: {}", str(results))
-        for r in results.titles:
-            if kind and kind != r.kind:
-                continue
-            if year and year != r.year:
-                continue
-            if r.title.lower() == movie_name.lower():
-                logger.debug("{} Matched {}", movie_name, r)
-                return_movie = r
-                break
-        # for/else hell yeah!
-        else:
-            logger.debug("{} Unmatched", movie_name)
-            return None
-
-    return return_movie
 
 
 def capitalize_movie_name(movie_name: str) -> str:
